fix: build fail from useSearchParams requiring Suspense boundary
